oncrpcsvc: inject request subject into processing thread
Motivation:
When a request is executed it can be processed by a number of components.
It's not always possible to pass caller's subject and, as a result, client
information get lost.

Modification:
Update RpcDispatcher to inject requestor's subject into current thread if
configured to do so. Update OncRpcSvc and OncRpcSvcBuilder to enable/disable
subject propagation. Update README with an example.

Result:
Callers subject available in the processing thread's AccessControlContext.

Acked-by: Gerd Behrmann
Target: master