xdr: rename org.dcache.utils.Opaque into org.dcache.xdr.XdrOpaque
Motivation:
Java9 module subsystem doesn't allow different one package to exists
in two jars files. As org.dcache.utils is too generic and exists in
other cache components as well, we have to avoid it's usage.

Modification:
move Opaque class into org.dcache.xdr package as XdrOpaque. In addition
the it implements XdrAble to be allow encoding/decoding if needed as
well as to be similar to other primitive types in org.dcache.xdr
package.

Result:
yet another set in direction of java9 module system.
NOTICE: this is a breaking change and existing clients of the library
require code changes.

Acked-by: Paul Millar
Target: master