rpc: introduce RpcTransport#isTLS
Motivation:
In some situations client or server wants to now if TLS in applied on
the connection, for example before sending the data over the wire or
for data protection validation.

Modification:
introduced  RpcTransport#isTLS that checks existence if SSLFilter in
the grizzly pipeline.

Result:
simple way to check TLS status of a connection.

target: master
Acked-by: Lea Morschel