core: Don't duplicate Buffers when not necessary
We're currently duplicate()-ing Buffers in two instances where we
read-ahead some data, so we don't have to revert to the old position
afterwards.

While it's a somewhat simple operation for regular HeapByteBuffers,
Buffer#duplicate() is not a free operation. Unfortunately, especially
with Grizzly's BuffersBuffer (or other implementations that we may use),
its use needs to be justified.

Replace the two instances with code that keeps track of the original
position before the read-access, reverting back to it afterwards,
without requiring new buffer allocations.

Also remove an explicit byteOrder(BIG_ENDIAN) call; all our buffers are
BIG_ENDIAN, just like the default.

Signed-off-by: Christian Kohlschütter <christian@kohlschutter.com>